EDMT-243 refactor : amplitude 유저 프로퍼티에 isAdmin 필드 삭제 및 props 리팩토링
