+++ src/components/student-record-write/student-record-write.tsx
@@ -74,7 +74,6 @@ export default function StudentRecordWrite({ recordType, recordId }: StudentReco
       <CharacteristicInput
         students={data}
         selectedId={parsedRecordId}
-        recordType={recordType}
         onGenerationStart={handleGenerationStart}
         onResponseGenerated={handleAiResponseGenerated}
       />
